pcsx2-gui: Cleanup my test variable names
I think one of the reasons I ran into this issue is because the second argument of FindWindowById is omitted in these lines.  The parent window can be provided which might scope things better (even if an id `0` exists somewhere).

Though it doesn't seem very simple to make this change.  I tried a few different ways and all got knocked down for one reason or another.  It works for now.. :shrug: