Added real CRCs for Jak protos
You are not supposed to run the DISKINFO.BIN or the FIREWIRE.IRX files directly. But due to the limitations of the pnach system, we must do it to get Debug Mode. So I decided that the easiest solution would be to add their CRCs here to fix their graphics. I wouldn't/couldn't add this change to official pcsx2, and it is not needed there anyway.